requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
numpy
scikit-learn
PyMuPDF
Pillow
transformers
torch